requirement.txt
amqp==2.6.1
anyjson==0.3.3
billiard==3.6.4.0
celery
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
django-celery
django-celery-beat==2.2.1
django-crispy-forms==1.13.0
django-formtools==2.2
django-import-export==2.4.0
django-picklefield==3.0.1
django-reversion==4.0.1
django-timezone-field==4.2.1
et-xmlfile==1.1.0
flower==1.0.0
future==0.18.2
httplib2==0.20.2
humanize==3.13.1
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
#prometheus-client==0.12.0
#prompt-toolkit==3.0.24
#pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
six==1.10.0
tablib==3.1.0
tornado==6.1
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
